requirements.txt
python-telegram-bot==13.15
requests==2.31.0
python-dotenv==1.0.0
flask==2.3.3
google-generativeai==0.3.1
psutil
matplotlib
folium
selenium
webdriver-manager
pillow
python-telegram-bot==13.7
requests
docx
pillow
jinja2
basemap
matplotlib
pandas
pandas
matplotlib
numpy
pillow
redis
pymemcache